Use dynamically constructed replacement for GDK_PIXBUF_LIBDIR on Windows
Instead of the hardcoded configure- and compile-time GDK_PIXBUF_LIBDIR
which is not expected to exist at run-time at all, use a pathname
constructed from where the program binary is, and assuming the normal
conventions.

This gets used in the --update-cache case to construct the pathname of
the loaders.cache file.

Note that there already is similar code in main() to construct the
pathname to where the loader DLLs normally are. There is probably some
degree of overlap there that could be factored out, but I don't care.