Bug 639922: Visual C++ 2008 Project Files
After checking with Tor Lillqvist, these are committed here...

They are done in a similar way to what is now in GLib and GTK+, so that
if the main GDK-Pixbuf library receives new source files, the new items
will be added to gdk-pixbuf.vcproj automatically when a tarball is released
via 'make dist'.

Note that the main GDK-Pixbuf library can be compiled in 2 variants, one
using the native GDI+ APIs from Windows which is supplied with the Windows
Platform SDK (but still makes use of LibPNG+ZLib) and another that relies
on 3rd-party image manipulation libraries (libTIFF, IJG JPEG and libJasPer
[JPEG-2000 library]) like what is used on other platforms AFAIK.  In either
case the image loaders are built directly into the main GDK-Pixbuf library
(i.e. the INCLUDE_<loader> macros are used).

This will build and "install" the main GDK-Pixbuf library and the
gdk-pixbuf-csource and gdk-pixbuf-query-loaders utility programs.