Fix handling of options when saving pngs
The png loader treats its options in two passes, the first pass
counts and verifies them, the second pass converts text options.
This worked ok when only text was understood, but with the introduction
of compression and icc-profile as supported options, the two loops
can now get out of sync when the text options are not all at the
beginning.

Fix this by skipping non-text options in the second loop, just
as we do in the first loop.

https://bugzilla.gnome.org/show_bug.cgi?id=620911