Unbreak jhbuild for gdk-pixbuf
The problem is pretty simple -- despite the guards against .po files
being updated on 'make' in the GNU gettext Makefile, if 'gdk-pixbuf.pot'
is newer than any .po file, the .po file will be updated from it.

git plays a strange game with timestamps, of course -- and this is the
source of our pain.  Even if gdk-pixbuf.pot was less-recently updated
than the .po files, you end up with the same timestamps on both (the
time of the checkout) and the .po gets remade.

We also have the problem that if the .pot file doesn't exist at all, it
is generated on the spot.

The only way to solve this issue is to have an old .pot file.  To that
end, we can just touch it to a date in the distant past from the
autogen.sh.

Since the content of the .pot file doesn't matter if it's old, we may as
well let touch create it as an empty file -- so let's remove it from
version control.

https://bugzilla.gnome.org/show_bug.cgi?id=661128