config.h.win32(.in): Enable USE_GMODULE
Enable USE_GMODULE for all builds, as one may want to build additional
image loader modules for use with GDK-Pixbuf later, to extend its
functionality.  For example, this is useful if one wants to build librsvg
with its GDK-Pixbuf image loader module, which is not included in the main
GDK-Pixbuf library.  The loader module DLLs need to be placed in
<parent_folder_of_GDK-Pixbuf_DLL>\lib\gdk-pixbuf-2.0\2.10.0\loaders, and
needs to be "registered" using the gdk-pixbuf-query-loaders utility
program.

Note: The image loaders that are included in GDK-Pixbuf continue to be
built in the main GDK-Pixbuf DLL for Visual C++ builds, in order to
simplify distribution of the library.