Make the tiff loader threadsafe
The reason why the tiff loader was marked as not threadsafe is
that libtiff only has a global error handler; while we still have
to set an error handler to keep libtiff from spamming stderr,
we just give up on collecting any error messages from it, and
we also give up on the illusion that we can push/pop error handlers.
That doesn't work without a stack, so we just keep forcing our
do-nothing error handler on without ever taking it back.