queryloaders: Ensure we link to libgobject
This call is necessary to ensure we actually link against libgobject;
otherwise it may be stripped if -Wl,--as-needed is in use.

The reason we need to link against libgobject is because it now has
a global constructor.  If the dynamically loaded modules happen
to dlclose() libgobject, then reopen it again, we're in for trouble.

See https://bugzilla.gnome.org/show_bug.cgi?id=686822

Patch-suggested-by: Ryan Lortie <desrt@desrt.ca>
Signed-off-by: Colin Walters <walters@verbum.org>

https://bugzilla.gnome.org/show_bug.cgi?id=686822