Fix parsing of JPEG orientation tag
Parsing the orientation tag in big-endian file formats got broken in
commit 0179dfd as reported in https://bugs.launchpad.net/bugs/1077186

A short (16 bit) value like the orientation value occupies the two first
bytes of the 4-byte offset field, independently of endianness. The
commit changed the code from using de_get16() to de_get32 to read out
this value, which broke on big-endian (MM/Motorola) byte order JPEG file
formats. I believe little-endian files would still work as long as the
two last, unused bytes were zero.

From http://www.cipa.jp/english/hyoujunka/kikaku/pdf/DC-008-2010_E.pdf

"Value Offset
This tag records the offset from the start of the TIFF header to the
position where the value itself is recorded. In cases where the value
fits in 4 Bytes, the value itself is recorded. If the value is smaller
than 4 Bytes, the value is stored in the 4-Byte area starting from the
left, i.e., from the lower end of the byte offset area. For example, in
big endian format, if the type is SHORT and the value is 1, it is
recorded as 00010000.H."

Also avoid the confusing recycling of the "offset" variable, use a
better name for the second use and make it and its siblings
local to the block where they are used.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>

https://bugzilla.gnome.org/show_bug.cgi?id=688427