Fix GIcon implementation
The "new rules" for GIcon say that we must support serialisation via
g_icon_serialize() and loadability via GLoadableIcon, so implement both
of those.

Serialise GdkPixbuf by emitting a GVariant that will result in a
png-encoded GBytesIcon when deserialized.  The GLoadableIcon interface
is similar: we return a stream that will read out as a png.

Test the serialisation by round-tripping an image through this process
and ensuring that it is pixel-perfect.

https://bugzilla.gnome.org/show_bug.cgi?id=688820