Make scaling of xpms work again
Commit c62676a284 had an unintended side-effect for loaders
which omit to call size_func, the new size set by
gdk_pixbuf_loader_set_size would overwrite the original
pixbuf size, and in effect force a scale factor of 1.0.
The xpm loader is one of the few which omit the size_func
call, thus the regression that calling
gdk_pixbuf_new_from_file_at_scale does not scale xpms anymore.
The fix is to use separate variables to pass the dimensions
when calling the size_func on behalf of the loader.

https://bugzilla.gnome.org/show_bug.cgi?id=686514