Avoid a discrepancy in file recognition
We are using a loader to implement some of the APIs that
are taking a filename, such as gdk_pixbuf_new_from_file_at_scale.
But the loader does not have the filename, so it can't use it
when determining the image type. This makes a difference for
types such as xbm, which have no good magic in the mime database.
Fix this by introducing a private API that lets us pass the
filename on to the loader.