Improve mime type detection
The mime database has no magic at all for xbm, and it
gets sniffed as text/plain. This is a regression from
when we were using our own magic, so be a little more
forgiving here - treat a mimetype of 'text/plain' like
uncertain results, and consult the filename. This lets
us at least recognize files with the extension .xbm
correctly again.

https://bugzilla.gnome.org/show_bug.cgi?id=573726