MSVC builds: Add NMake Makefile for Introspection Builds
This adds an NMake Makefile that is used to generate the introspection
files for GDK-Pixbuf, which replaces the previous approach of generating
them via the Visual Studio Projects.  This makes it clearer to people on
what things are needed in order to build the introspection files if so
desired, and makes things a bit cleaner as the build results can be
cleaned away easily.  This also takes the recent change in g-i into
account, where we no longer pass in the DLL file name but instead pass in
the library name on Windows, just as it was done on *nix.

The existing .bat file for this purpose is removed as a result.