MSVC Introspection Builds: Clean up a bit
Since gobject-introsection-1.40.0 we are passing in the .lib file(s) to
deduce the DLL(s) that the introspection files should look for, instead of
just passing in the DLL file(s), we no longer need to make a copy of the
main .lib file to match the <namespace-name>-<namespace-version> convention
for use during the linking stage of the introspection build.  Remove that
from the introspection builds as a result.