lib: Remove incorrect info about area-prepared signal
It is not safe to fill the pixbuf returned from area-prepared with a
background colour as a number of loaders (notably XPM and SVG) will
only have information about the size and characteristics of the image
after it's been fully loaded. Filling the background colour would then
overwrite the image we just loaded.

Given that this won't work properly with anything but 100% transparent
or 100% solid pixels, remove that advice from the docs. Instead,
applications should rely on their toolkits to composite over a
background colour.

https://bugzilla.gnome.org/show_bug.cgi?id=589334