MSVC 2008 Builds: Clean Up and Speed Up
Make the builds that use the open source imaging libraries the default
Debug and Release builds, and make the GDI+ builds have configs suffixed
with _GDI+, as the builds using the open source imaging libraries are in
a much better state, at least on more modern versions of Windows.

Also standardize on the linked libraries for all builds, regardless whether
a release or debug build is used, no more 'd' suffixes in the .lib
filenames, and move the .lib's to link into the property sheets, to ease
future maintenance.

Speed up release builds by using /MP for the compiler options, which can
cut down on release build times by quite a bit.  This does emit a brief
warning for debug builds, but the code would otherwise build normally.