MSVC Builds: "Install" the .pdb Files
"Install" the .pdb files for all tools and DLL that are built, so that it
would make it easier to debug builds, especially when developing the GTK+
stack or so, and we already generate the .pdb files for all builds.

Also be more selective on the files that we copy, to ensure that we only
copy the executables and DLLs and LIBs that are part of gdk-pixbuf, so that
we avoid copying items that are not meant to be copied or copy plugin DLLs
incorrectly when the projects are included from an all-in-one solution used
to build the entire GTK+ stack, for example.