Clean Up Visual Studio Project Generation
This makes use of the common autotools modules that was just added so that
there would need to be less items in gdk-pixbuf's main Makefile.am, so that
we can generate the complete Visual Studio 2008/2010 project files (which
will then be used to obtain the Visual Studio 2012 and 2013 projects).

The other added benefit to this is that the property sheets used to install
headers will also make use of this mechanism so that the listing of public
headers can be kept up-to-date whenever the list of headers et updated.