gif: Fix thumbnailing GIF animations
When using progressive loading, the GIF loader seems to expect us to
load every single frame of the image before being "done". That's really
not the case when thumbnailing, as we only want the first frame in that
case.

Return the GDK_PIXBUF_ERROR_INCOMPLETE_ANIMATION error when not all the
frames have been decoded, so that users of the API can ignore it, like
gnome-desktop's thumbnailer.

https://bugzilla.gnome.org/show_bug.cgi?id=755672