MSVC builds: Move math.h compatibility header
Move the math.h compatibility header from build/win32/math-compat to
build/win32/vs9, and copy it to build/win32/vs10 and build/win32/vs11
during 'make dist'.  This is because we don't need the header for
Visual Studio 2013 and later, and the math.h header shipped with
Visual Studio 2015 is now included in Windows 10 SDK's UCRT directory,
so the old way of including the compiler-shipped math.h by doing
'#include <../include/math.h>' would not work there.

This will fix the build on Visual Studio 2015.

https://bugzilla.gnome.org/show_bug.cgi?id=756381