Visual Studio builds: Have package version available in projects
This is to make generating .pc files for the package easier for Visual
Studio builds.  Also make GlibEtcInstallRoot more consistent across the
board, and add setting entries for installation path to the Python
interpretor, also to make things easier for generating .pc files, as well
as building the introspection files.