Fix bug 766842
This commit fixes an issue where pixops_composite_nearest_noscale () had
different behavior than pixops_composite_nearest () for pixels in the
destination that are not a part of the transformed source image. Per the
documentation, the pixels at the edges of the source image should be
replicated to infinity.

Added a test to tests/pixbuf-composite.c

https://bugzilla.gnome.org/show_bug.cgi?id=766842