Add version macro header
Add a version macro header that would reveal the availability of the
various APIs, which is used to  indicate what stable release series an
API is available, and what release series is an API deprecated.  This
is also used to export the symbols using compiler directives, which
would eliminate the need to maintain .symbols files.

Annotate the symbols in the pubic headers accordingly, and update the
doc sections file so that the macros are included in the documentation.

https://bugzilla.gnome.org/show_bug.cgi?id=767164