gdk-pixbuf: Use compiler directives to export symbols
Like what is now done in GLib, ATK, GTK+ and some other libraries, make use
of compiler directives that is determined at configure stage (or
pre-configured) to export the symbols for the GdkPixbuf library.  This
cleans up things a bit, as the .symbols file will no longer need to
be maintained, as long as the GDK_PIXBUF_AVAILABLE_IN_* (or so)
annotation is marked in the header and the source file that defines
the function has config.h included first.

Clean up configure.ac and gdk-pixbuf/Makefile.am a bit as a result.

https://bugzilla.gnome.org/show_bug.cgi?id=767164