pixdata: Avoid copying more data than available
The problem we were encountering in bug 775693 wasn't that we were
running past the end of the memory we just allocated, but that the
length of the data we were given couldn't possibly match the dimensions
of the image.

This fixes running past the end of the given pixdata when the length is
provided inside the structure.

https://bugzilla.gnome.org/show_bug.cgi?id=775693