ico: Fix uninitialised alpha values
As we always create a GdkPixbuf with alpha, even if the original file
doesn't have an alpha channel, or a mask, make sure to set the alpha
pixel value in all cases, if only to have it overridden if there is a
mask present.

This makes it possible to use the reference test without having
differing alpha values.