io-ico: Fix option parsing to be locale-independent
sscanf() and strtol() are both locale-dependent. In addition, the return
value of sscanf() was not being checked (so it could fail without being
noticed), and there was no bounds checking being performed.

Bounds checking is now performed, although the bounds have been chosen
for conservative backwards-compatibility, and may not be the most
appropriate.

Coverity CID: 1388522

https://bugzilla.gnome.org/show_bug.cgi?id=776990