thumbnailer: Update from gnome-thumbnailer-skeleton master
https://github.com/hadess/gnome-thumbnailer-skeleton/commits/master

Add some assertions to help Coverity

From looking at the code, these assertions should always hold already,
but it clarifies the code and helps static analysis if we add them
explicitly.

(n_pixels == 0) is only true if (s_x1 == s_x2 || s_y1 == s_y2), which is
only true if (dx == 0 || dy == 0). This cannot be true, because dx and
dy are the result of dividing source_[width|height] by
dest_[width|height], and there’s already an assertion that the former
are greater or equal to the latter.

Coverity CID: 1388528

https://bugzilla.gnome.org/show_bug.cgi?id=776990