gdk-pixbuf-loader: Add a missing NULL check
The code just above checks whether image_module is NULL, and it doesn’t
look like it can change in the meantime, so we should also check before
dereferencing it for load_increment.

Make this an assertion; although we’re not entirely sure how this code
was originally supposed to work, we’re still in development, so we can
investigate any assertion failures with impunity.

Coverity CID: 1388529

https://bugzilla.gnome.org/show_bug.cgi?id=776990