gdk-pixbuf-io: Add precondition checks for image dimensions
Refuse to save an invalid GdkPixbuf: one which has negative width,
height or number of channels.

This allows us to put assertions for all those properties in the I/O
code, which helps static analysis not go off on one because it thinks
the image width could be negative.

Coverity ID: 1388534

https://bugzilla.gnome.org/show_bug.cgi?id=777374