Visual Studio builds: Move projects to win32/
This moves the Visual Studio project files and related items under win32/
so that people can go one less directory layer down to the Visual Studio
projects.  This also makes things in sync with what is now on GLib and GTK+,
and cleans up the Visual Studio 201x projects as some items can actually be
combined.