Visual Studio builds: Support Visual Studio 2017
Update the autotools scripts so that we can support Visual Studio 2017 by
copying the 2010 projects and updating items as necessary.

Note that the toolset version format for 2017 has changed, so we now
allow a custom string to be set and thus be used; otherwise the toolset
version string is generated as it was before.

Also note that Visual Studio 2017 aims to be compatible with Visual Studio
2015 at the CRT level, meaning that one should be able to use 2017-compiled
binaries with 2015-compiled binaries without problems.