Builds: Add a fallback-c89.c for pre-C99 compilers
Some supported compilers lack some C99 math functions that are beginning
to be used in the code, so this adds a fallback implementation for them,
namely round() and lrint(), where the fallbacks are used if they are not
found during configure.

For Visual Studio builds, config.h.win32.in is updated to reflect the
situation on various Visual Studio versions as Visual Studio builds
do not use autotools.