Initial implementation of a Meson build
Meson is a better build system than Autotools that has gotten traction
in the GNOME community.

The build is pretty much equivalent between Autotools and Meson, but at
this point it needs further testing to verify complete matching
behaviour.

The GDI-plus loaders are not currently built, as that will require
access to a platform I don't have in order to test them.

Additionally, the API reference is not being generated yet in Meson.