build: Fix running gen-thumbnailer.py on Windows
We need to prepend the PATH envvar so that the main GDK-Pixbuf DLL that
we just built can be loaded on Windows when running this script, as
Windows do not support rpath.

Since gdk-pixbuf-pixdata resides in the same path as the main GDK-Pixbuf
DLL, we can just use its path so that we can acquire the needed path to it.

https://bugzilla.gnome.org/show_bug.cgi?id=785767