build: Force-include msvc_recommended_pragmas.h on MSVC
Since we are always building against GLib, make use of the
msvc_recommended_pragmas.h that comes with it, to clean things up a bit,
as that will help us to silence unwanted build warnings as well as
raising attention for warnings that could be potential code issues.
This is the header that is force-included in the Visual Studio projects.

See
https://git.gnome.org/browse/glib/commit/?id=5ba3b4022ebda9e1756d568b7baac7f285bfc0b4 for its use in building GLib.

https://bugzilla.gnome.org/show_bug.cgi?id=785767