meson: Use configure_file() to copy typesfile
The working directory is the builddir and gnome.gtkdoc is not smart
enough to figure out whether the filename(s) passed are found in the
builddir or srcdir. Using files() doesn't help, either, since it always
generates paths to the builddir.

Other projects (GStreamer, Grilo) seem to use configure_file to copy the
typesfile to the builddir. Follow this.

This fixes the following error:

    Cannot open gdk-pixbuf.types: No such file or directory

https://bugzilla.gnome.org/show_bug.cgi?id=785964