build: Add fallback dependency discovery for MSVC on Meson
Some of the external dependencies we use for loaders do not generate
pkg-config files within their Visual Studio build system, namely libpng
and libtiff.  So, we need to add a fallback mechanism so that they can
be correctly found in Visual Studio builds when their pkg-config file(s)
could not be found, by using cc.has_header() and cc.find_library(), so
that one can avoid needing to craft pkg-config files for them, which can
be error-prone.

For both of these cases, since they both support DLL + import .lib and
static .lib builds, we look for the DLL + import .lib first, and then
fall back to the static builds.

https://bugzilla.gnome.org/show_bug.cgi?id=785767