build: Allow different Python installations per MSVC 201x projects
Make the Python release series used default to the following:

-Python 3.4.x for Visual Studio 2010, 2012 and 2013
-Python 3.6.x for Visual Studio 2015 and 2017

This is to make it in-line with what is used in GObject-Introspection,
as an MSVC subproject will be added shortly to enable building the
introspection files within the Visual Studio projects, so that we can
be more certain that we use the same Python installation that was used
to build GObject-Introspection.  The defaults are selected in this
manner as the official CPython binaries on Windows are built with 2010
(3.4.x) and 2015 (3.6.x) respectively.

Note that the 2008 projects are not changed here as Python 2.7.x is the
only CPython version series that is built with 2008 and supports the
things in G-I fully, so the 2008 projects continues to default to a
Python 2.7.x installation.