build: Support introspection builds in MSVC projects
This integrates the NMake Makefiles into the MSVC projects for building
GDK-Pixbuf, to make the build process easier.  Note that the
introspection files are not built by default, but must be explicitly
built and cleaned.

This will require a complete G-I build in $(GLibEtcInstallRoot) with the
Python installation that was used to build G-I in order for this to
work.

Update the build documentation for the MSVC projects to reflect on this
enhancement.