build: Support Windows components-based loader
There is currently a GDI+ image loader that is within the GDK-Pixbuf
codebase, which can serve as a basis for improvements to it or for
moving on to more modern Windows APIs/Frameworks such as Windows
Imaging Component (WIC).

Provide a Windows-only option to build it with Meson, which will mean
that when it is used the TIFF and JPEG loaders will not require libtiff
and libjpeg-turbo/IJG JPEG,  but will use the GDI+ component of Windows
to deal with these images, along with BMP, GIF, ICO and images.  This
will also enable the support of EMF and WMF images.  For building this
set of loaders into the GDK-Pixbuf DLL, pass in 'windows' as a part of
the loaders passed into builtin_loaders, or pass in 'all' to
builtin_loaders in conjunction with this option.

This also means that by using this set of GDI+ loaders, we cannot use
any of the loaders (either in-house or those using the various
opensource libraries) for these formats, so these loaders will be
disabled as a result, since the MIME types, extensions and signature
of these formats are the same in the GDI+ loaders with their respective
counterparts that are otherwise built.

Note that PNG support still requires libpng, and JPEG2000 support still
requires libjasper, and in its current state, the TIFF and JPEG and ICO
(and hence ANI) support that are done by the opensource libraries and/or
the in-house loaders are still preferred over the GDI+-based ones since
in the current state they tend to work better, so this is provided as an
option.

https://bugzilla.gnome.org/show_bug.cgi?id=785767