tests: Bump timeouts for long-running tests
The pixbuf-area-updated and pixbuf-randomly-modified tests are
particularly taxing on the CPU, and though modern desktop machines
shouldn't have a problem dealing with those under 30 seconds, tests
are run in parallel, and busy build machines might fail those tests
because of timeouts.

Bump the timeouts to 300 seconds for those long tests to avoid test
failures.