MSVC projects: Improve binary security
Ensure that the /DYNAMICBASE linker option is enabled (which is the
default setting), as we inadvertly disabled it in our projects
explicitly, to improve security of our MSVC-built binaries.

For MSVC 2012 and later, use /HIGHENTROPYVA for linking x64 builds to
enhance security of the binaries as well.

Pointed out by Ignacio Casal Quinteiro.