build: Update the introspection rules
We should not be using cross-compilation to gate the generation of
introspection data; it's perfectly possible to use a helper binary
to run g-ir-scanner when cross-compiling — in fact, that's what
projects like Yocto already do.

We should also return a target for the introspection data, to allow
using gdk-pixbuf as a sub-project.