loader: Expose the dimensions of the original image
Currently, it's not possible to get the dimensions of the original
image from a scaled GdkPixbuf. This is problematic for thumbnailers
because they want to embed the original dimensions into the resulting
thumbnail file. The only way to get this information is to either
decode the full-resolution image and then downscale it, even when the
decoder supports downscaling on-the-fly; or to re-implement large
chunks of code to avoid decoding the full-resolution image and
retain the original dimensions.

Exposing the original dimensions as options simplifies the thumbnailer
without affecting performance.

https://bugzilla.gnome.org/show_bug.cgi?id=778517