Silence two compiler warnings
We have two "set but unused variable" warnings coming from
gdk-pixbuf-xlib. The first is caused by an unnecessary variable used to
hold the return value of XGetWindowAttributes(); we don't do anything
with it, and it's way too late for changing the behaviour of the
function to bail out on error. The second is caused by a variable that
is only used when a debugging printf() is compiled in; to avoid the
warning without creating even more complex macro hell, we tell the
compiler that the variable is unused.