GdkPixbuf: Formalize the difference between pixels/bytes storage
Instead of checking the bytes vs. pixels fields all over the place, we
introduce an enum that says which mode the pixbuf is in.  We have a
central downgrade_to_pixels() function to turn the pixbuf from
immutable to mutable.

This commit only does the gdk-pixbuf core; we'll deal with pixdata and
the loaders separately.