Generate GdkPixdata introspection
The GdkPixdata API is built as part of the GdkPixbuf shared library,
but:

 - it has its own namespace
 - it has its own header file
 - it's mostly meant for C applications
 - it is deprecated

Consumers of the introspection data cannot access this API by using the
GdkPixbuf-2.0 GIR, as the namespace and included headers are different.

Instead of kludging the API, let's move the GdkPixdata introspection
out of the GdkPixbuf-2.0 GIR and into its own module.

This is an introspection ABI break, but there aren't many consumers of
GdkPixdata out there, and it's preferable to straight including the
gdk-pixdata.h header into gdk-pixbuf.h.

Fixes #72