jasper: Avoid a deprecation warning
The jas_stream_memopen() function uses any non-negative buffer size with
a NULL buffer to signify that the buffer will be allocated internally
and set to grow as needed. In the future, jas_stream_memopen() will be
replaced by a function that takes a size_t instead of a signed integer.

Passing -1 as the buffer size trips a run time deprecation warning, so
we can just use 0 and keep compatibility between the old libjasper and
newer versions.

Closes: #73