build: Add post install script for MSVC builds
Visual Studio builds do not assume the presence of a shell script
interpreter but would normally use cmd.exe, which would make the post
install shell script useless in such situations.

Fix this by adding a Windows .bat port of the post install script, to be
used on Visual Studio builds, which may be usable on MinGW builds if
MinGW builds can be done directly in cmd.exe.