Speed up saving PNGs by skipping a null transform
Calling png_set_shift to set the number of significant bits of input
triggers a transform operation inside libpng which shifts bits around,
even if it's a null operation of 8 bits to 8 bits.

I measured this slowdown as about 30% of total saving runtime on a
large 7680x2160 screenshot.

In addition, only 8 bits per channel are supported in gdk-pixbuf so
this is never actually needed in practice...

Skipping the png_set_shift call altogether doesn't change output and
speeds it up nicely.

Fixes #90